+++ Aleksandrova_Vilena/2/checker/checker.rb
@@ -1,10 +1,9 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'optparse'
 require 'logger'
 
-require_relative 'lib/helper'
+require_relative 'lib/args_parser'
 require_relative 'lib/ping'
 
 options = {}
