+++ Aleksandrova_Vilena/2/checker/lib/logging.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+require 'logger'
+
+# Logger
+module Logging
+  class << self
+    def logger
+      @logger ||= Logger.new($stdout)
+    end
+
+    attr_writer :logger
+  end
+
+  def self.included(base)
+    class << base
+      def logger
+        Logging.logger
+      end
+    end
+  end
+
+  def logger
+    Logging.logger
+  end
+end
